+++ apps/next/pages/banner.tsx
@@ -71,6 +71,7 @@ import { bannerPresets } from '@app/modules/banner/bannerPresets';
 import { BannerPresetList } from '@app/modules/banner/BannerPresetList';
 import Layout from '@app/components/layout';
 import { ChangePresetModal } from '@app/modules/banner/ChangePresetModal';
+import { usePaymentPlan } from '@app/util/hooks/usePaymentPlan';
 
 const bannerEndpoint = '/api/features/banner';
 const defaultForeground: keyof typeof BannerForegrounds = 'ImLive';
@@ -162,6 +163,8 @@ const bannerTypes = {
         settings: emggBannerSettings,
     },
 };
+
+const lockedBanners = ['Emgg', 'Sway'];
 
 interface Props {
     banner: Banner;
@@ -264,9 +267,7 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
 
 export default function Page({ banner, originalBanner }: Props) {
     const { data: sessionInfo } = useSession();
-
-    const { data: paymentPlanResponse } = useSWR<APIPaymentObject>('payment', async () => (await fetch('/api/user/subscription')).json());
-    const paymentPlan: PaymentPlan = paymentPlanResponse === undefined ? 'Free' : paymentPlanResponse.plan;
+    const [paymentPlan, paymentPlanResponse] = usePaymentPlan();
 
     const { data: streamingState } = useSWR('streamState', async () => await (await fetch(`/api/twitch/streaming/${session?.user['id']}`)).json());
     const streaming = streamingState ? streamingState.isStreaming : false;
@@ -281,7 +282,7 @@ export default function Page({ banner, originalBanner }: Props) {
 
     const applyPreset = (preset: BannerPresetProps) => {
         setFgId(preset.foreground.id);
-        setFgProps(preset.foreground.props);
+        setFgProps({ ...preset.foreground.props, ...((banner?.foregroundProps as any)?.username ? { username: (banner?.foregroundProps as any)?.username } : {}) });
         setBgId(preset.background.id);
         setBgProps(preset.background.props);
     };
@@ -290,6 +291,7 @@ export default function Page({ banner, originalBanner }: Props) {
 
     useEffect(() => {
         if (preset && preset !== 'custom') {
+            setTabIndex(0);
             applyPreset(bannerPresets[preset as keyof typeof bannerPresets] ?? defaultBanner);
         }
     }, [preset]);
@@ -379,6 +381,8 @@ export default function Page({ banner, originalBanner }: Props) {
     const { isOpen: pricingIsOpen, onOpen: pricingOnOpen, onClose: pricingClose, onToggle: pricingToggle } = useDisclosure();
     const { isOpen: disableBannerIsOpen, onClose: disableBannerOnClose, onToggle: bannerDisabledToggle } = useDisclosure();
     const [sliderValue, setSliderValue] = useState(0);
+
+    const [tabIndex, setTabIndex] = useState(0);
 
     useEffect(() => {
         const getSliderValue = () => {
@@ -552,10 +556,10 @@ export default function Page({ banner, originalBanner }: Props) {
                         </Center>
 
                         <Flex {...styles} grow={1} p="4" my="4" rounded="md" w="full" direction="column" minH="lg">
-                            <Tabs colorScheme="purple" flexGrow={1} size={breakpoint !== 'base' ? 'md' : 'sm'}>
+                            <Tabs colorScheme="purple" flexGrow={1} size={breakpoint !== 'base' ? 'md' : 'sm'} index={tabIndex} onChange={setTabIndex}>
                                 <TabList>
                                     <Tab className={trackEvent('click', 'banner-tab')}>Banner</Tab>
-                                    {preset !== 'emgg' && <Tab className={trackEvent('click', 'background-tab')}>Background</Tab>}
+                                    <Tab className={trackEvent('click', 'background-tab')}>Background</Tab>
                                 </TabList>
 
                                 <TabPanels flexGrow={1}>
@@ -637,27 +641,26 @@ export default function Page({ banner, originalBanner }: Props) {
                                         </VStack>
                                     </TabPanel>
                                     <TabPanel>
-                                        <FormControl id="country">
-                                            <FormLabel>Background type</FormLabel>
-                                            <RadioGroup
-                                                onChange={(e) => {
-                                                    setBgId(e as keyof typeof BackgroundTemplates);
-                                                }}
-                                                value={bgId}
-                                            >
-                                                <Stack direction="column">
-                                                    {Object.entries(BackgroundTemplates).map(([key, value]) => (
-                                                        <Radio key={key} value={key} colorScheme="purple">
-                                                            <Text fontSize="md">{value.name}</Text>
-                                                            <Text fontSize="sm">
-                                                                {value.description}
-                                                            </Text>
-                                                        </Radio>
-                                                    ))}
-                                                </Stack>
-                                            </RadioGroup>
-                                        </FormControl>
-                                        {fgId !== 'Emgg' && (
+                                    {!lockedBanners.includes(fgId) ? (
+                                        <>
+                                            <FormControl id="country">
+                                                <FormLabel>Background type</FormLabel>
+                                                <RadioGroup
+                                                    onChange={(e) => {
+                                                        setBgId(e as keyof typeof BackgroundTemplates);
+                                                    }}
+                                                    value={bgId}
+                                                >
+                                                    <Stack direction="column">
+                                                        {Object.entries(BackgroundTemplates).map(([key, value]) => (
+                                                            <Radio key={key} value={key} colorScheme="purple">
+                                                                <Text fontSize="md">{value.name}</Text>
+                                                                <Text fontSize="sm">{value.description}</Text>
+                                                            </Radio>
+                                                        ))}
+                                                    </Stack>
+                                                </RadioGroup>
+                                            </FormControl>
                                             <Box py="4">
                                                 <Form
                                                     setProps={(p) => {
@@ -668,8 +671,11 @@ export default function Page({ banner, originalBanner }: Props) {
                                                     accountLevel={paymentPlan}
                                                 />
                                             </Box>
-                                        )}
-                                    </TabPanel>
+                                        </>
+                                    ) : (
+                                        <Text pt='2'>You cannot customize the background of this banner template. Choose another template to customize the background.</Text>
+                                    )}
+                                        </TabPanel>
                                 </TabPanels>
                             </Tabs>