+++ apps/next/pages/pricing.tsx
@@ -26,29 +26,27 @@ import {
     Flex,
     Link,
     Box,
-    Stack,
-    List,
-    ListIcon,
-    ListItem,
-    Spacer,
     WrapItem,
 } from '@chakra-ui/react';
 
 import getStripe from '../util/getStripe';
 import prisma from '../util/ssr/prisma';
-import { FaTwitter, FaCheck, FaArrowRight } from 'react-icons/fa';
+import { FaTwitter, FaCheck } from 'react-icons/fa';
 import { ProductCard } from '@app/components/pricing/ProductCard';
 import { trackEvent } from '@app/util/umami/trackEvent';
 import { APIPaymentObject, PaymentPlan } from '@app/util/database/paymentHelpers';
+import { NextSeo } from 'next-seo';
+import { generalFaqItems, pricingFaqItems } from '@app/modules/faq/data';
+import { FaqSection } from '@app/modules/faq/FaqSection';
+import { usePaymentPlan } from '@app/util/hooks/usePaymentPlan';
+import { FreeProductCard } from '@app/components/pricing/FreeProductCard';
 
 type Props = {
     products: (Product & { prices: Price[] })[];
 };
 
 const Page: NextPage<Props> = ({ products }) => {
-    console.log('products: ', products);
-    const [paymentPlan, setPaymentPlan] = useState<PaymentPlan>();
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const [paymentPlan, paymentPlanResponse] = usePaymentPlan();
     const { status, data: session } = useSession({ required: false }) as any;
 
     const router = useRouter();
@@ -82,20 +80,6 @@ const Page: NextPage<Props> = ({ products }) => {
         })[]
     ) => products.filter((a) => a.name.includes('Gift'));
 
-    useEffect(() => {
-        (async () => {
-            if (status === 'authenticated') {
-                const res = await fetch('/api/user/subscription');
-
-                const data: APIPaymentObject = await res.json();
-
-                if (data) {
-                    setPaymentPlan(data.plan);
-                }
-            }
-        })();
-    }, [status]);
-
     const handlePricingClick = useCallback(
         async (priceId: string, isSubscription: boolean) => {
             router.push({
@@ -104,9 +88,16 @@ const Page: NextPage<Props> = ({ products }) => {
                 },
             });
             if (ensureSignUp()) {
-                // TODO - this needs to change
-                if (paymentPlan === 'Professional') {
-                    return router.push('/account');
+                if (paymentPlan === 'Professional' || paymentPlan === 'Personal') {
+                    const res = await fetch('/api/stripe/create-portal', {
+                        method: 'POST',
+                        headers: {
+                            'content-type': 'application/json',
+                        },
+                    });
+
+                    const data = await res.json();
+                    router.push(data.subscriptionUrl);
                 }
 
                 const res = await fetch('/api/stripe/create-checkout-session', {
@@ -186,185 +177,39 @@ const Page: NextPage<Props> = ({ products }) => {
         <>
             {test()}
             <Button>test</Button>
-        </>
+
+            <VStack spacing={[6, 12]} w="full">
+                {AnnualBillingControl}
+                <Center w={['auto', 'auto', 'auto', '5xl']}>
+                    <SimpleGrid columns={[1, 1, 1, 3]} spacing="4" w="full">
+                        <WrapItem key={'free2'} w="full" h="full">
+                            <FreeProductCard />
+                        </WrapItem>
+                        {sortProductsByPrice(products).map((product) => (
+                            <ProductCard
+                                key={product.id}
+                                product={product}
+                                billingInterval={billingInterval}
+                                handlePricingClick={(priceId) => handlePricingClick(priceId, true)}
+                                paymentPlan={paymentPlan}
+                                paymentPlanResponse={paymentPlanResponse}
+                            />
+                        ))}
+                    </SimpleGrid>
+                </Center>
+                <Container centerContent maxW="container.lg" experimental_spaceY="4">
+                    <Text fontSize="md">Prices in USD. VAT may apply. Membership is tied to one Twitter account.</Text>
+                    <Text textAlign="center" maxW="2xl" px="4" fontSize="xl">
+                        Just like you, the people behind PulseBanner are creators. And like you, we rely on PulseBanner Memberships to keep improving and maintaining PulseBanner.
+                        Supporting PulseBanner enables us to do what we love and empower creators ❤️
+                    </Text>
+                    <Box pt="8">
+                        <FaqSection items={pricingFaqItems.concat(generalFaqItems)} />
+                    </Box>
+                </Container>
+            </VStack>
+            </>
     );
-
-    // return (
-    //     <>
-    //         <NextSeo title="Pricing" />
-    //         <Modal onClose={onClose} size={'xl'} isOpen={isOpen}>
-    //             <ModalOverlay />
-    //             <ModalContent>
-    //                 <ModalHeader>
-    //                     <Center>Almost there!</Center>
-    //                     <Center>Connect to Twitter to continue.</Center>
-    //                 </ModalHeader>
-    //                 <ModalCloseButton />
-    //                 <ModalBody minH="32" h="32" pb="4">
-    //                     <Flex h="full" direction="column" justifyContent="space-between">
-    //                         <VStack grow={1}>
-    //                             <Button
-    //                                 onClick={() => {
-    //                                     if (session?.accounts?.twitter) {
-    //                                         return;
-    //                                     }
-    //                                     const url = new window.URL(window.location.href);
-    //                                     url.searchParams.append('modal', 'true');
-
-    //                                     signIn('twitter', {
-    //                                         callbackUrl: url.pathname + url.search,
-    //                                     });
-    //                                 }}
-    //                                 colorScheme="twitter"
-    //                                 leftIcon={<FaTwitter />}
-    //                                 rightIcon={session?.accounts?.twitter ? <FaCheck /> : undefined}
-    //                             >
-    //                                 Connect to Twitter
-    //                             </Button>
-    //                         </VStack>
-    //                         <Center>
-    //                             <Text fontSize="sm">
-    //                                 {'By signing up, you agree to our'}{' '}
-    //                                 <Link as={NextLink} href="/terms" passHref>
-    //                                     Terms
-    //                                 </Link>{' '}
-    //                                 and{' '}
-    //                                 <Link as={NextLink} href="/privacy" passHref>
-    //                                     Privacy Policy
-    //                                 </Link>
-    //                             </Text>
-    //                         </Center>
-    //                     </Flex>
-    //                 </ModalBody>
-    //             </ModalContent>
-    //         </Modal>
-
-    //         <Container maxW="container.lg" experimental_spaceY="6" pb="6">
-    //             <Heading size="xl" textAlign="center" h="full">
-    //                 PulseBanner Memberships
-    //             </Heading>
-    //             <Center>
-    //                 <Text textAlign="center" fontSize="xl" maxW="container.sm">
-    //                     You can use PulseBanner for free forever 🎉 OR you can unlock even more awesome features and kindly support the creators with a PulseBanner Membership.
-    //                 </Text>
-    //             </Center>
-    //         </Container>
-    //         <VStack spacing={[6, 12]} w="full">
-    //             {AnnualBillingControl}
-    //             <HStack>
-    //                 {giftingProducts(products).forEach((product) => {
-    //                     product.prices.map((price) => {
-    //                         console.log('product info: ', price.id);
-    //                         return (
-    //                             <Button onClick={() => handlePricingClick(price.id)} key={price.id}>
-    //                                 {product.name}
-    //                             </Button>
-    //                         );
-    //                     });
-    //                 })}
-    //                 ;
-    //             </HStack>
-    //             <Center w={['auto', 'auto', 'auto', '5xl']}>
-    //                 <SimpleGrid columns={[1, 1, 1, 3]} spacing="4" w="full">
-    //                     <WrapItem key="free" w="full" h="full">
-    //                         <Card props={{ w: 'full', h: 'full' }}>
-    //                             <Box w="full" experimental_spaceY={4}>
-    //                                 <Flex direction="row" justify="space-between" alignItems="center">
-    //                                     <VStack alignItems="start" spacing={0}>
-    //                                         <Heading size="lg">Free</Heading>
-    //                                         <Text>Features with limited customization</Text>
-    //                                     </VStack>
-    //                                 </Flex>
-    //                             </Box>
-    //                             <Flex direction="row" justify="space-between" alignItems="center" justifyContent="center">
-    //                                 <VStack spacing={0} cursor="pointer">
-    //                                     <Stack direction={['column', 'row']} alignItems={['center', 'center']} w="full" spacing={[0, 2]}>
-    //                                         <Text
-    //                                             fontSize="2xl"
-    //                                             fontWeight="extrabold"
-    //                                             lineHeight="tight"
-    //                                             as={chakra.span}
-    //                                             bg="green.200"
-    //                                             px="1"
-    //                                             py="0.5"
-    //                                             mb="4"
-    //                                             rounded="md"
-    //                                             color="black"
-    //                                         >
-    //                                             Free
-    //                                         </Text>
-    //                                     </Stack>
-    //                                 </VStack>
-    //                             </Flex>
-
-    //                             <Box flexGrow={2} experimental_spaceY={2}>
-    //                                 <Heading size="md">{"What's included"}</Heading>
-    //                                 <List>
-    //                                     {['Twitter Live Banner', 'Twitter Name Changer'].map((feature) => (
-    //                                         <ListItem key={feature}>
-    //                                             <ListIcon color="green.300" as={CheckIcon} />
-    //                                             {feature}
-    //                                         </ListItem>
-    //                                     ))}
-    //                                 </List>
-    //                                 <Heading size="md">{'What am I missing?'}</Heading>
-    //                                 <List>
-    //                                     <ListItem key="profile image">
-    //                                         <ListIcon color="red.400" as={CloseIcon} />
-    //                                         Live Twitter Profile Picture
-    //                                     </ListItem>
-    //                                     <ListItem key="profile image">
-    //                                         <ListIcon color="red.400" as={CloseIcon} />
-    //                                         Banner refreshing
-    //                                     </ListItem>
-    //                                     <ListItem key="profile image">
-    //                                         <ListIcon color="red.400" as={CloseIcon} />
-    //                                         Custom banner background image
-    //                                     </ListItem>
-    //                                 </List>
-    //                             </Box>
-
-    //                             <Box justifySelf="flex-end">
-    //                                 <Flex w="full" justifyContent="space-between">
-    //                                     <Spacer />
-    //                                     {session && (
-    //                                         <Button
-    //                                             fontWeight="bold"
-    //                                             colorScheme="green"
-    //                                             rightIcon={<FaArrowRight />}
-    //                                             onClick={() => {
-    //                                                 const url = new window.URL(window.location.href);
-    //                                                 url.searchParams.append('modal', 'true');
-    //                                                 signIn('twitter', {
-    //                                                     callbackUrl: url.pathname + url.search,
-    //                                                 });
-    //                                             }}
-    //                                         >
-    //                                             Sign up
-    //                                         </Button>
-    //                                     )}
-    //                                 </Flex>
-    //                             </Box>
-    //                         </Card>
-    //                     </WrapItem>
-    //                     {sortProductsByPrice(products).map((product) => (
-    //                         <ProductCard key={product.id} product={product} billingInterval={billingInterval} handlePricingClick={handlePricingClick} />
-    //                     ))}
-    //                 </SimpleGrid>
-    //             </Center>
-    //             <Container centerContent maxW="container.lg" experimental_spaceY="4">
-    //                 <Text fontSize="md">Prices in USD. VAT may apply. Membership is tied to one Twitter account.</Text>
-    //                 <Text textAlign="center" maxW="2xl" px="4" fontSize="xl">
-    //                     Just like you, the people behind PulseBanner are creators. And like you, we rely on PulseBanner Memberships to keep improving and maintaining PulseBanner.
-    //                     Supporting PulseBanner enables us to do what we love and empower creators ❤️
-    //                 </Text>
-    //                 <Box pt="8">
-    //                     <FaqSection items={pricingFaqItems.concat(generalFaqItems)} />
-    //                 </Box>
-    //             </Container>
-    //         </VStack>
-    //     </>
-    // );
 };
 
 // Since we export getServerSideProps method in this file, it means this page will be rendered on the server